Fix pub dev linter
Pub dev is not allowing to publish because of README. There was an a
piece of text left from the template and pub dev recognizes as an
unwritten readme.

Uploading... (1.7s)
`README.md` contains a generic text fragment coming from package templates (`TODO: Include short and useful examples for package users`).
Please follow the guides to write great package pages:
https://dart.dev/guides/libraries/writing-package-pages